Raise exception if optimization for plain contents fails
Since the new optimization for plain contents seeks immediately
to the next dollar sign at the beginning of each iteration of the
parser's main loop, we should never encounter any character other
than a dollar sign at the beginning of the list when parsing plain
contents. Raising an exception if we do encounter a symbol other
than a dollar sign should help debugging any future issues that
might arise in relation to this optimization.