Fix bug in if parsing code
This fixes a bug introduced in commit f67f40581. As part of breaking
up the code into multiple files, that commit rewrote Str and Var
to be initialized with plain strings, rather than with PlainToken
and MetaToken objects. I did forget to update that handling for the
implicit else arm for if statements without an explicit else clause.

This commit fixes that bug. It also introduces a new test case to
catch any possible future regressions.