Bump version to 1.0.2.dev
Having another look at semver.org, it seems that the `dev`-suffix
is supposed to be used for denoting a pre-release. as such, it would
be sorted before, not after the corresponding number. Hence, when
bumping the version after a release, I should probably already bump
the version number to the next release I anticipate when adding the
`dev`-suffix. Also note that setuptools encourages the use of a dot
as separator before the `dev` suffix, while it suggests using a minus
sign for post-releases.